dunnlikewarren

+++ buglog.client/src/components/Note.vue
@@ -1,11 +1,15 @@
 <template>
-  <div class="col-4 card mt-1 bg-darkgrey">
+  <div class="col-4 card my-1 bg-darkgrey text-color">
     <div class="card-body">
-      <p class="text-color">
+      <h5>
+        Note:
+      </h5>
+      <p>
         {{ note.body }} <span class="float-right cursor-pointer" @click="deleteNote"><i class="fa fa-trash text-light" aria-hidden="true"></i></span>
       </p>
+      <img :src="note.creator ? note.creator.picture : 'https://st4.depositphotos.com/4329009/19956/v/600/depositphotos_199564354-stock-illustration-creative-vector-illustration-default-avatar.jpg'" alt="">
       <p>
-        Author: {{ }}
+        Author: {{ note.creator.name }}
       </p>
     </div>
   </div>
@@ -27,7 +31,9 @@ export default {
     return {
       state,
       deleteNote() {
-        noteService.deleteNote(props.note.id, props.note.bug)
+        if (window.confirm('Are You Sure?')) {
+          noteService.deleteNote(props.note.id, props.note.bug)
+        }
       }
     }
   },
